range: don't truncate length to size_t bits

+++ loader/common/range.h
@@ -11,7 +11,7 @@ void range_align_start(struct range *range, size_t alignment);
 void range_align_length(struct range *range, size_t alignment);
 void range_align(struct range *range, size_t alignment);
 
-static inline void range_advance_begin(struct range *range, size_t by)
+static inline void range_advance_begin(struct range *range, u64 by)
 {
     range->begin += by;
 }
@@ -31,12 +31,12 @@ static inline bool range_is_empty(const struct range *range)
     return (range->end - range->begin) == 0;
 }
 
-static inline void range_set_length(struct range *range, size_t length)
+static inline void range_set_length(struct range *range, u64 length)
 {
     range->end = range->begin + length;
 }
 
-static inline size_t range_length(struct range *range)
+static inline u64 range_length(struct range *range)
 {
     return range->end - range->begin;
 }
